requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
playwright>=1.44.0
structlog>=24.1.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiofiles>=23.2.1
aiohttp>=3.9.0
